EMPTY: BUG: ClutterInputDeviceTool not shared among tablets
It appears that each tablet has its own list of ClutterInputDeviceTool objects.
If the same tool is used on multiple tablets, a caller will recieve distinct
ClutterInputDeviceTool objects that share the same type and serial. Its debatable
whether this is desirable, but does cause issues for the Wayland backend of
Mutter, since clients are informed per-tablet about a single tool when clients
are supposed to only be informed per-seat.

See signature of _clutter_input_device_lookup_tool